Integrate the CRA templates from NPM
