+++ src/constants/styles.ts
@@ -1,0 +1,7 @@
+export const ROOT_COMPONENT = {
+    angular: '<body class="pxb-blue mat-typography mat-app-background">',
+    ionic: '<app-root class="pxb-blue mat-typography mat-app-background"></app-root>',
+    reactNative: '',
+};
+
+export const STYLES = 'html, body {\r\n\theight: 100%;\r\n\tmargin: 0;\r\n\tpadding: 0;\r\n}';
