+++ src/extensions/pxblue-extensions.ts
@@ -11,12 +11,9 @@ import {
     SCRIPTS,
     ROOT_COMPONENT,
     STYLES,
-    ROOT_IMPORTS,
     PRETTIER_DEPENDENCIES,
     PRETTIER_SCRIPTS,
     PRETTIER_CONFIG,
-    APP_COMPONENT,
-    QUESTIONS,
 } from '../constants';
 import {
     updateScripts,
@@ -178,26 +175,10 @@ module.exports = (toolbox: GluegunToolbox): void => {
     };
 
     const addPXBlueReact = async (props: ReactProps): Promise<void> => {
-        const { name, lint, prettier, language, template } = props;
+        const { name, lint, prettier, language } = props;
         const folder = `./${name}`;
         const ts = language === 'ts';
         const isYarn = filesystem.exists(`./${folder}/yarn.lock`) === 'file';
-
-        // Install Dependencies
-        await fileModify.installDependencies({
-            folder: folder,
-            dependencies: DEPENDENCIES.react,
-            dev: false,
-            description: 'PX Blue React Dependencies',
-        });
-
-        // Install DevDependencies
-        await fileModify.installDependencies({
-            folder: folder,
-            dependencies: DEV_DEPENDENCIES.react,
-            dev: true,
-            description: 'PX Blue React Dev Dependencies',
-        });
 
         // Install ESLint Packages (optional)
         if (ts && lint) {
@@ -211,10 +192,6 @@ module.exports = (toolbox: GluegunToolbox): void => {
                 folder: folder,
                 config: LINT_CONFIG.tsx,
             });
-
-            let webVitals = filesystem.read(`${folder}/src/reportWebVitals.ts`);
-            webVitals = `/* eslint-disable */\r\n${webVitals}`;
-            filesystem.write(`${folder}/src/reportWebVitals.ts`, webVitals);
         }
 
         // Install Code Formatting Packages (optional)
@@ -227,83 +204,16 @@ module.exports = (toolbox: GluegunToolbox): void => {
             });
         }
 
-        // Final Steps: browser support, styles, theme integration
+        // Final Steps: lint and prettier scripts (if applicable) and app title
         const spinner = print.spin('Performing some final cleanup...');
 
         // Update package.json
         let packageJSON: any = filesystem.read(`${folder}/package.json`, 'json');
-        packageJSON = updateScripts(packageJSON, SCRIPTS.react.concat(lint && ts ? LINT_SCRIPTS.react : []));
-        packageJSON = updateScripts(packageJSON, SCRIPTS.react.concat(prettier ? PRETTIER_SCRIPTS.react : []));
+        packageJSON = updateScripts(packageJSON, lint && ts ? LINT_SCRIPTS.react : []);
+        packageJSON = updateScripts(packageJSON, prettier ? PRETTIER_SCRIPTS.react : []);
         packageJSON = updateBrowsersListJson(packageJSON);
         if (prettier) packageJSON.prettier = '@pxblue/prettier-config';
         filesystem.write(`${folder}/package.json`, packageJSON, { jsonIndent: 4 });
-
-        // Update index.css
-        filesystem.remove(`${folder}/src/index.css`);
-        filesystem.write(`${folder}/src/index.css`, STYLES);
-
-        // Update index.js/tsx
-        let index = filesystem.read(`${folder}/src/index.${!ts ? 'js' : 'tsx'}`, 'utf8');
-        const imports = ROOT_IMPORTS.react.join('\r\n');
-        index = `import 'react-app-polyfill/ie11';\r\nimport 'react-app-polyfill/stable';\r\n${index}`
-            .replace('ReactDOM.render(', `${imports}\r\n\r\nReactDOM.render(`)
-            .replace('<App />', ROOT_COMPONENT.react);
-        filesystem.write(`${folder}/src/index.${!ts ? 'js' : 'tsx'}`, index);
-
-        if (ts && lint) {
-            // update the App.tsx to pass linting
-            let app = filesystem.read(`${folder}/src/App.${!ts ? 'js' : 'tsx'}`, 'utf8');
-            app = APP_COMPONENT.react;
-            filesystem.write(`${folder}/src/App.${!ts ? 'js' : 'tsx'}`, app);
-        }
-
-        // If the user opt in for templates
-        if (template.length !== 0) {
-            // @ts-ignore
-            const templateOptions: string[] = QUESTIONS.template.choices.map((choice) => choice.name);
-
-            // clear App.css
-            filesystem.write(`${folder}/src/App.css`, '');
-
-            // replace index.css
-            toolbox.template.generate({
-                template: 'react/index.css.ejs',
-                target: `${folder}/src/index.css`,
-            });
-
-            // if user opts in for dashboard
-            if (template.includes(templateOptions[2])) {
-                // replace App.tsx
-                toolbox.template.generate({
-                    template: 'react/App.tsx.ejs',
-                    target: `${folder}/src/App.${ts ? 'tsx' : 'js'}`,
-                    props: {
-                        ts,
-                    },
-                });
-
-                // replace App.test.tsx
-                toolbox.template.generate({
-                    template: 'react/App.test.tsx.ejs',
-                    target: `${folder}/src/App.test.${ts ? 'tsx' : 'js'}`,
-                    props: {
-                        ts,
-                    },
-                });
-
-                // add src/pages/dashboard.tsx
-                toolbox.template.generate({
-                    template: 'react/pages/dashboard.tsx.ejs',
-                    target: `${folder}/src/pages/dashboard.${ts ? 'tsx' : 'js'}`,
-                    props: {
-                        ts,
-                    },
-                });
-
-                // remove unused logo.svg
-                filesystem.remove(`${folder}/src/logo.svg`);
-            }
-        }
 
         // Update index.html
         let html = filesystem.read(`${folder}/public/index.html`, 'utf8');
