--- src/templates/react/App.test.tsx.ejs
@@ -1,9 +1,0 @@
-import React from 'react';
-import { render, screen } from '@testing-library/react';
-import App from './App';
-
-test('renders app bar with the correct text', ()<%= props.ts ? ': void' : '' %> => {
-    render(<App />);
-    const linkElement = screen.getByText(/Dashboard/i);
-    expect(linkElement).toBeInTheDocument();
-});
