--- src/templates/react/App.tsx.ejs
@@ -1,6 +1,0 @@
-import React from 'react';
-import DashboardPage from './pages/dashboard';
-
-const App<%= props.ts? ': React.FC' : '' %> = () => <DashboardPage />;
-
-export default App;
