--- src/templates/react/index.css.ejs
@@ -1,11 +1,0 @@
-html,
-body,
-#root {
-    margin: 0;
-    padding: 0;
-    height: 100%;
-    width: 100%;
-}
-* {
-    box-sizing: border-box;
-}
