+++ src/utilities/types.ts
@@ -1,6 +1,8 @@
 export type Language = 'ts' | 'js';
 export type Cli = 'expo' | 'rnc';
 
+export type Template = 'Blank' | 'Basic Routing' | 'Authentication';
+
 export type AngularProps = {
     name: string;
     lint: boolean;
@@ -11,7 +13,6 @@ export type ReactProps = {
     language: Language;
     lint: boolean;
     prettier: boolean;
-    template: string[];
 };
 export type ReactNativeProps = {
     name: string;
